% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iar_mle.R
\name{iar_mle}
\alias{iar_mle}
\title{Perform Maximum Likelihood Estimation for the Infinite Autoregressive (IAR) model}
\usage{
iar_mle(par = 0, fn = iar_loglik, data, hessian = TRUE, method = "Brent")
}
\arguments{
\item{par}{The initial parameter value for optimization. Default is 0.}

\item{fn}{The likelihood function to be optimized. Default is \code{iar_loglik}.}

\item{data}{The input data for the likelihood function.}

\item{hessian}{A logical value indicating whether to compute the Hessian matrix. Default is TRUE.}

\item{method}{The optimization method to be used. Default is "Brent".}
}
\value{
The optimized parameter values.
}
\description{
This function performs maximum likelihood estimation for the IAR model using the specified likelihood function and optimization method.
}
\examples{
data <- c(1, 2, 3, 4, 5)
optimized_params <- iar.mle(data = data)

}
\keyword{IAR}
\keyword{estimation,}
\keyword{likelihood}
\keyword{maximum}
\keyword{model,}
\keyword{optimization}
