% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iar_sim.R
\name{iar_sim}
\alias{iar_sim}
\title{Simulate a time series using the Irregular Autoregressive (IAR) model}
\usage{
iar_sim(t_n, phi = 0.7, sigma = 1)
}
\arguments{
\item{t_n}{A vector representing the time points.}

\item{phi}{The autoregressive parameter. Default is 0.7.}

\item{sigma}{The standard deviation of the error term. Default is 1.}
}
\value{
A data frame containing the simulated time series values and corresponding time points.
}
\description{
This function generates a simulated time series using the IAR model.
}
\examples{
t_n <- 1:10
sim_data <- iar.sim(t_n)

}
\keyword{IAR}
\keyword{model}
\keyword{series,}
\keyword{simulation,}
\keyword{time}
