% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_its}
\alias{plot_its}
\title{Plot function}
\usage{
plot_its(x, tn)
}
\arguments{
\item{x}{A vector of x values.}

\item{tn}{A vector of tn values.}
}
\description{
This function creates a line plot of x versus tn and adds two custom axes.
}
\examples{
# Generate some data
x <- rnorm(100)
tn <- seq(1, 100)
# Use the function
plot.its(x, tn)
}
